Don't be mad, but I created this
This is a working checklist model. It is dynamically sized and can add or remove items and clear the whole list. I just threw the entire program in here because I don't know what you need to make it work on your systems. 

Also, I quickly made a logo on MS Paint that can definitely be updated.....