Login and New Account Registration Pages
I've got the LogIn page and the Account Registration form working. There is a bit of code that I've commented out since it will need to be implemented with the rest of the program and have the name changed to match everything up (Basically, where it says "MainForm()" will need to be changed, and then lines 28, 31, 54, 148-183, 198-204, 214-268 will need to be uncommented). Also, for SQLite, I created a database on my computer that won't exist on yours, so the database path on line 62 must be changed to whatever Canyon has set up.
Those changes are only for the LogIn.cs code.  The RegisterForm.cs code is perfect.